postinst: fix the removal of all extensions
When upgrading to a new version, if there are two or more extensions
enabled, all extensions are deleted due to incorrect verification logic.
Only the last downloaded extension remains. This makes it impossible
to boot the system after a reboot in closed environments due to lack of
access to https://update.release.flatcar-linux.net

Signed-off-by: Sergey Garaev <sergeygaraev2019@gmail.com>